Uses Terraform for host certificates
TL;DR
-----

Creates certificates for vSphere hosts with Terraform

Details
-------

Uses Terraform to create certificates for the vSphere hosts with
Let's Encrypt. Assumes that the certs are built from FQDN CSRs
generated from the ESXi GUI. Each CSR should be placed in the
`secrets` directory at the repository root in a `${HOST.csr}`.