Creates a certificate for HAProxy
TL;DR
-----

Prepares for K8s install by generating a certificate for HAProxy

Details
-------

vSphere with Tanzu requires an HAProxy instance when not using NSX.
The HAProxy appliance that's recommended requires a TLS cert as
part of installing the OVA. This update creates a certificate for
the HAProxy instance.

I'm still not 100% on why the certificate is needed since HAProxy
will balance request for Kubernetes Ingress endpoints that are
providing their own certificates. Nevertheless I'm making sure
that I have a legit cetificate as part of the installation.